(#5782) - Handle rev-1 conflict during sync
* (#5782) - add sync test for fuzzy failure

Adds a test to reproduce the error detected in #5782

Failing test for fuzzy

minimal repro

* (#5782) - fetch rev-1 conflicts in replication

We currently have an optimisation that uses _bulk_docs to fetch
rev-1 documents during replication. This reduces the number of requests
to a remote server by avoiding individual document GETs (though this
is now mitigated by the _bulk_get endpoint in CouchDB 2.0), but can
only be used for documents that have exactly one revision because we
are not fetching the complete rev tree.

This patch fixes an issue with the optimisation where we would fail to
fetch a conflicting revision at rev-1. For instance, if 2 documents
with identical _ids but different _revs were created on the local and
remote datastores.

The medium term solution is to remove the optimisation (#4651) in
favour of _bulk_get, but this should allow us to maintain backwards
compatibility until _bulk_get is widely deployed.