fix(memleak): remove the destruction listener on close()
Calling `close()` is a final method, the same way that calling `destroy()` is a final method.
Even if another instance closes the same database, there is no reason we should notify our instance's user of it (since it notified its intent to be done with our instance by expressely calling `close()`).