(#5793) - fixes for sync against Couchbase Sync Gateway
* (#5793) - test _conflicts: [] is rejected by bulk_docs

* (#5793) - test _conflicts in allDocs

allDocs should not contain a _conflicts array
when no conflicts exist. This is the behaviour in CouchDB but
not in other implementations.

* (#5793) - do not return empty _conflicts array

Previously, PouchDB would include an empty conflicts array in the
response to a document fetch with conflicts=true specified. This patch
adjusts it to match the CouchDB behaviour, omitting the _conflicts
field altogether if there are no conflicts for a given document.

* (#5793) - do not POST _conflicts during sync

Couchbase Sync Gateway throws errors when a document within a bulk_docs
request includes a _conflicts field. Strictly speaking, this is likely
a CSG error - it should behave the same as CouchDB - but this is a
pragmatic fix to keep PouchDB compatible.