fix(destruct-listeners): prevent race conditions
This fix allows `close()` to actively destroy the destructListeners array even if the events are out-of-sync.
It also moves all the magic of dealing with the destruct listeners arrays into `setup.js` and provides APIs for the events that may deal with it.
(I'm not 100% sure providing an event API rather than plain `.ref()` and `.unref()` functions is the best way to do this; nice thing is that it allows to listen on those events as well which might be useful to test suites?)

Notice that `destructiListeners` now stores PouchDB instances rather than closures.

Tests were added to try to catch possible mistakes.