(#5857) - support latest=true for GET /doc
This adds support for the latest=true option when fetching a document.
The associated tests best describe the expected behaviour (based on
what CouchDB does) but, essentially, it enables a user to pass an old
revision in the `rev` or `open_revs` parameter and have the current
associated leaf revision returned.

This is primarily used by replicators, because a document may be updated
/compacted between the `_changes` feed being read and the resulting
document fetch, would would result in a "missing" revision. Using
`latest=true` guarantees that the doc GET will always return a result.