Update version of debug dependency
TLDR; The 2.3.2 version of debug wipes localStorage on load when built using browserify.

Extended remix: Including 6.1.0 of pouchdb in a browserified build causes debug to lose its localStorage setting on load.

Cause: The load() method of the 2.3.2 version of debug doesn't return immediately after getting localStorage. It then checks for `process.env`. This evaluates `true` in a browserified environment, causing the localStorage value to be dropped. This is then persisted to localStorage.

The 2.6.0 version returns immediately on getting the localStorage value, so the existing value is re-persisted.