(#5717) - Support selector filter in changes()
As a precursor to supporting selectors in replication, add support for
filtering _changes using a selector. Note that this doesn't work in
CouchDB 1.x which predates the find/selector feature.

The bulk of this commit involves splitting out the core selector logic
from the pouchdb-find plugin into a new module, pouchdb-selector-core,
that can be used in the pouchdb core without pulling in all of pouchdb-find.
There are no unit tests for the new module - we rely on the integration tests
for changes() and pouchdb-find.

The istanbul ignores can be removed when PouchDB server has this functionality,
and so the tests can run over the http adapter as well.

fixup