(#5717) - Support selector filter for replication
Building on the _changes support for selector, this allows a selector
filter to be specified during replication.

Checkpoint behaviour has been updated to account for selectors, accepting
that depending on JSON.stringify may be unreliable due to instability
in the implementation.

The istanbul/ignores could be removed when pouch-server supports
selector-filtered changes/replication.