pouchdb-adapter-idb: fix "possible EventEmitter memory leak detected"
Problem
-------

With the idb adapter (default in Chrome), calling `.changes()` or
`.replicate()` with `{live: true}` multiple times results in this error,
even when the handler is nicely terminated with `.cancel()`.

    (node) warning: possible EventEmitter memory leak detected. 11
    listeners added. Use emitter.setMaxListeners() to increase limit.

Related to #4176, #4141, #4093, #3848, #2667, #2311, #2205, and maybe
others.

How to reproduce the bug?
-------------------------

Run this plunkr: http://plnkr.co/edit/JKCbQ6vwlYClAp6ZSAf4?p=preview

Alternatively, you can create + cancel a `.changes()` feed a few times:

    changesHandler = db.changes({live: true});
    changesHandler.cancel();
    repeat

About this fix
--------------

The bug comes from the `_change()` API call from the idb adapter [1]
returning `undefined` instead of the `{cancel: function() {...}}` object
it should return [2]. Because of that, the cleanup function in the core
module doesn't call the adapter's `cancel()` function [3], letting a
listener into the wild.

This commit fixes that and adds a test to check for the presence of the
error log line.

[1]: https://github.com/pouchdb/pouchdb/blob/49e4945/packages/node_modules/pouchdb-adapter-idb/src/index.js#L420-L422
[2]: https://github.com/pouchdb/pouchdb/blob/49e4945/packages/node_modules/pouchdb-adapter-idb/src/changes.js#L29-L33
[3]: https://github.com/pouchdb/pouchdb/blob/49e4945/packages/node_modules/pouchdb-core/src/changes.js#L201-L209