(#6510) - validate _changes timing parameters
PouchDB already has logic to create a buffer between the server side
timeout and the client timeout, providing a 5 second grace period
for the server to respond before the client will timeout. However, there
are some situations where this causes _changes to fail with a client
timeout. Specifically:

1. If the timeout is less than 5 seconds. Given the logic to add a 5
second grace period, we need the client request timeout to be at least
this.
2. If the heartbeat is specified that is greater than the client timeout.
If a heartbeat parameter is specified, Couch ignores any timeout parameter.

To handle (1), this commit adjusts the client side timeout to be at least
the timeout + the buffer period (5 seconds). To handle (2), we make the
client side timeout at least the heartbeat value + the buffer period.