(#7095) - Reduce replication to server requests
This solves an issue where a bulkDocs request of two docs with a
continuous replication to the server causes the docs to be sent
in two replication requests rather than one.

Firstly this modifies the runBatchedCursor implementation to use
getAll() and getAllKeys() when a batchSize of -1 (or infinite) is
given. This means the onBatch callback will fire once instead of
once per doc.

Secondly this changes the idb changes implementation to call the
onChange handler in the same tick so all changes get added to the
pending batch before it's processed.