Deterministic revision numbers with md5
This uses md5 to create the the revision number for a document. This
means that each document will have a deterministic revision nubmer.
This makes client side conflict resolution easier and will reduce the
number of conflicts

By default this is turned on, but a user can fallback to generating
random revision numbers by passing `useMd5Rev: false' as an option to a
new PouchDB instance