Making openTransactionSafely safe for idbnext, some error guarding
Using _openTransactionSafely in the api should force everyone to go
through the same idb reference, meaning if a migration is occuring it
will block and get you a correct reference. Additionally, asking for a
reference to be closed allows existing transactions to finish, so it
should work in both directions.

I also added some guards to attempt to surface setup errors to callers,
though it's not perfect. I think to make this better we might need to
alter the API layer in a large way.