Rewriting booleans and null to be indexable
This somewhat changes what doc.data means. Previously it ws used for
indexing and in some places shorthand for the most recent version when
getting.

Now it is only for indexing, and is deep cloned and modified to deal
with differences between what IndexedDB supports and what CouchDB
requires. To load the real submitted value pull from doc.revs[doc.rev]
instead.