(#7582) - Stream test results and beautify output
One of the difficult things I encountered when I first worked on PouchDB
was to understand the output of the remote (i.e. selenium and saucelabs)
tests on Travis.

This PR enhances the test output for selenium/saucelabs test jobs. When
the test is remote (selenium/saucelabs), the mocha runner events are sent
over the network and a mocha Spec reporter is used locally. The output at
the end of the tests cleanly show the failed tests, their actual/expected
values of the failed assertions and their errors stack traces.

The changes have been tested at home:
- with npm run dev (the reporter is still the one in the browser)
- with selenium
- with saucelabs
- with BAIL=0
- with PERF=1
- manually adding errors to test cases to test bailing