(#6543) - Implement pouchdb-find natively for IDBNext
* Fast, native indexes for find + IDBNext

Previously find indexes were implemented by piggy-backing onto mapreduce
views, and generating a view for the index.

We now use native indexeddb indexes instead, which are much faster and
should be smaller on disk. They are maintained whenever the database is
opened, as well as when a find query is attempted for a native index
that does not yet exist.

This is supported over multiple browser tabs and sessions, in all the
major browsers, with the exclusion of IE11.

It is *nearly* completely compatible with CouchDB, with only two
minor differences (found so far):

1. Results from ranged queries that contain very small numbers may not
be ordered completely correctly. This is fixable, but it probably
carries an overhead unless we're quite careful, and probably will not
come up in practice.

2. IndexedDB does not index objects at all (ie if your key path points
to more JSON as opposed to a primitive type or an array that document's
entry will not exist in the index). This is not solvable, but it seems
unlikely that anyone would really use that feature.

There are tests for both of these issues in test.type.js.

NB: For future IDBNext development you should now always use $t to get a
transaction instead of any other method (eg directly asking idb for
one), as this process goes through the correct process to make sure you
always safely get a transaction handle regardless of what other execution
flows / tabs are doing.

NB: This release is not completely backwards compatible with existing
IDBNext databases, as there is no migration for the compatibility
cloning that occurs on write to allow for some index features that
CouchDB supports but IndexedDB does not. It will still run, but indexes
may be wrong.

* Dropping the number of iterations

These can take a lot of time, and we don't want Travis to kill our build
over it.