Optimising rewrite by performing a much faster first pass
Reading is so much faster than writing, and many documents will never
be hit by these rewrite rules.

So instead of rewriting everything each time, do preliminary checks to
see if a document needs to be rewritten.

Further potential optimisation might be to merge the test with the
rewrite and only rewrite the part of the tree that requires it (up to
the root).