(#7618) - Making the indexeddb adapter publicly consumable
* Remove the private flag to we publish this adapter to npm

* Stopping idb and indexeddb from opening each other

idb can't open indexeddb dbs because the version is too old. We detect
that and generate a useful error to the user, which happens when you
try to instantiate the handle (new PouchDB(..))

for indexeddb we check to see if the old version is in the range that
idb occupies but indexeddb never does. Unfortunately due to indexeddb
lazily opening handles this only happens once you actually perform
an action on the DB (dbhandle.post({foo: 'bar'}))

* Removing NEXT generation

* Apparently my editor's eslint plugin is broken again..

* Update travis to also test indexeddb adapter

* Does changing all chromes to saucelabs magically work with no additional effort?

* Dropping chrome_bin

* More debugging on selenium / sauce connections

* Removing unneeded logging