Fix warning no index found message
The warning message describing you are querying on a field that has no matching index set, has changed interpunction since CouchDB 3.1.
This commit with adhere to same interpunction (. at end of sentence, and starting with capital N instead of n).
The test in pouchdb-test depend on the string. So test would break.
Perhaps should add some code for backwards compatability with CouchDB < 3.1?