fix: _design/ docs are included in the limit
When querying the db , the result will include the store its design documents. So when passing a limit of say 1, and the other documents have an id of > '_design' (assuming a default alphabetic sorting of the primary key), the _design docs will be return first, then filtered. This results in 0 documents returned.A workaround for this (in this commit) is querying the total number of design docs, and adding that number to the limit passed to the query. Then processing the results later removing the number of design docs from total_rows in an attempt to be opaque about the hack.