Fix/repare fix (broken) tests
There are a couple of tests in pouchdb-find that use the No matching index found warning. We need to support both variations (CouchDB pre 3.1 and post 3.1).

I did not change the tests themselves. Most of them are still skipped, or not even part of the describe scope, therefor not run.

The pouchdb-find test suite and code still needs work, imo. But that's for another PR.