Fix testrunners of couchdb v2 and add couchdb v3
The warning message describing you are querying on a field that has no matching index set, has changed interpunction since CouchDB 3.1.
This commit with adhere to same interpunction (. at end of sentence, and starting with capital N instead of n).
The test in pouchdb-test depend on the string. So test would break.
Perhaps should add some code for backwards compatibility with CouchDB < 3.1?

There are a couple of tests in pouchdb-find that use the No matching index found warning. We need to support both variations (CouchDB pre 3.1 and post 3.1).

I did not change the tests themselves. Most of them are still skipped, or not even part of the describe scope, therefore not run.

The pouchdb-find test suite and code still needs work, imo. But that's for another PR.