Let caller set the change batch size when updating views (#8320)
* Refactor: pass batch size into `updateViewInQueue()`

We'd like to make the batch size for this process configurable by the
caller of the `query()` method. In order to make this possible, we
replace references to the `CHANGES_BATCH_SIZE` constant with an option
setting that's passed from `queryPromised()` into `updateView()` and
then into `updateViewInQueue()`.

* Let caller set the change batch size when updating views

Some users need to be able to increase or reduce the size of changes
batches used when updating view indexes, to make the best use of the
storage adapter they're running.

For example, memory constraints or processing capacity might necessitate
using smaller batch sizes.

This change lets the caller set a DB option named
`view_update_changes_batch_size`, which changes the batch size we use
for updating views in `abstract-mapreduce`. It defaults to 50, the value
of the existing `CHANGES_BATCH_SIZE` constant in this module.

* Document the `view_update_changes_batch_size` option