chore: remove obsolete replication revision-one optimization (#8499)
* chore: remove obsolete replication revision-one optimization

With #993, we introduced an optimization to retrieve documents via
`_all_docs` if they had a revision of 1, since there was no `_bulk_get`
API at that time. This optimization allowed us to retrieve many
revision-one documents with a single request, instead of requesting each
document individually. Documents with higher revisions still had to be
fetched one by one.

Since then, `_bulk_get` has been implemented in CouchDB and is used by
PouchDB's replicator, so this optimization is no longer necessary. On
the contrary, it now leads to additional requests that slow down
replication, although not that much.

* chore: remove left over log statement in tests

* fix: replication cancellation race condition

This fixes a race condition that was observed in a presumed flaky test
that was ultimately not flaky: If a replication is aborted at the exact
time when a checkpoint was written at the end of a changes batch without
changes, then the replication was not really aborted.