tests/performance: fix use of this.iterations in setup/tearDown
Currently, setting ITERATIONS to anything except 0 will silently break
`pull-replication-perf-skimdb` because of a mismatch between this.iterations in
the test `setup()` function and the actual number of iterations the test is run
for.

This failure is surfaced by https://github.com/pouchdb/pouchdb/pull/8768.

This commit moves access to iterations from a testCase's `this` to an argument
in the setup() function.