pouchdb-utils: fix nextTick "Illegal invocation" error (#9092)
* pouchdb-utils: fix nextTick "Illegal invocation" error

When instantiating a new PouchDB instance using the leveldb adapter in
an Electron renderer process, `nextTick()` is called with a different
context than is expected, resulting in `TypeError: Illegal invocation`.
Fix this by binding `queueMicrotask` to the `globalThis` context before
returning it.

* pouchdb-utils: bind undefined vs globalThis in nextTick

The linter does not recognize `globalThis`.  Use `undefined` instead,
which still fixes the `Illegal invocation` error occurring in Electron.