Fix test.issue915.js on a clean environment
The test attempts to create the file
`./tmp/_pouch_veryimportantfiles/something`, and by doing so implicitly
assumes that the directory `./tmp/_pouch_veryimportantfiles` already
exists, which may not be true in a clean environment.

This patch creates the directory with `{recursive: true}`, meaning that
if the directory already exists it will just succeed silently without
doing anything, and if it doesn't, the directory will be created.

Fixes #9120